Replace search_log with dump_log.
This will be a much more useful debugging tool.
